requirements.txt
attrs==19.3.0
boto3==1.14.14
botocore==1.17.14
docutils==0.15.2
et-xmlfile==1.0.1
jdcal==1.4.1
jmespath==0.10.0
more-itertools==8.4.0
numpy==1.19.0
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
redis==3.0.1
redis-py-cluster==2.0.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
XlsxWriter==1.2.9
dnspython==2.0.0
